pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "animus"
version = "0.1.0"
description = "Local-first AI agent with RAG and tool use"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
native = [
    "llama-cpp-python>=0.2",
]
embeddings = [
    "sentence-transformers>=2.0",
]
chromadb = [
    "chromadb>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "pytest-asyncio>=0.23",
]
all = [
    "animus[native,embeddings,chromadb,dev]",
]

[project.scripts]
animus = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
timeout = 30
testpaths = ["tests"]
addopts = "-v --tb=short"
